refactor(warehouse-sources): split external data source view
Generated-By: PostHog Desktop
Task-Id: 574f49f3-f788-4bbf-907f-b54af3716bac
