+++ products/warehouse_sources/backend/presentation/views/external_data_source/__init__.py
@@ -1,0 +1,1 @@
+"""external_data_source for views. One module per concern."""
